dockerfiles: Fix example dockerfile and provide an example start script
The existing dockerfile is no longer able to run HaasoscopeQt.py due
to recent Python 3.11 changes and due to module changes in the
HaasoscopeQt.py code.  Update the docker build rules to account for
those changes.

Also, add in a new haasoscope.sh example script with details on how to
run the container.  Update the documentation with a description of
that script.